# crm-dashboard/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  description: |
    API for managing church members, cell groups (CGF), attendance, ministries, and events.
    This API provides CRUD operations for all entities and dashboard analytics endpoints.
  version: 1.0.0
  contact:
    name: Church Admin
    email: admin@church.org
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.church.org/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /analytics/members/distribution:
    get:
      tags: [Analytics]
      summary: Get member distribution
      description: Get member distribution by various dimensions
      parameters:
        - name: type
          in: query
          required: true
          description: Distribution type
          schema:
            type: string
            enum: [gender, domicile, age_group, status]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        label:
                          type: string
                          example: "Laki-laki"
                        value:
                          type: integer
                          example: 250
                        percentage:
                          type: number
                          format: float
                          example: 50.0
              example:
                success: true
                data:
                  - label: "Laki-laki"
                    value: 250
                    percentage: 50.0
                  - label: "Perempuan"
                    value: 250
                    percentage: 50.0
  /analytics/members/trends:
    get:
      tags: [Analytics]
      summary: Get member growth trends
      description: Get monthly/weekly member growth data
      parameters:
        - name: period
          in: query
          description: Time period
          schema:
            type: string
            enum: [weekly, monthly, yearly]
            default: monthly
        - name: months
          in: query
          description: Number of months to include
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        period:
                          type: string
                          example: "2024-01"
                        new_members:
                          type: integer
                          example: 15
                        total_members:
                          type: integer
                          example: 485
                        active_members:
                          type: integer
                          example: 460
              example:
                success: true
                data:
                  - period: "2024-01"
                    new_members: 15
                    total_members: 485
                    active_members: 460
                  - period: "2024-02"
                    new_members: 12
                    total_members: 497
                    active_members: 470
  /analytics/attendance/summary:
    get:
      tags: [Analytics]
      summary: Get attendance summary
      description: Get attendance statistics and trends
      parameters:
        - name: cg_id
          in: query
          description: Filter by CGF ID
          schema:
            type: string
        - name: start_date
          in: query
          description: Start date
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_records:
                        type: integer
                        example: 1000
                      hadir:
                        type: integer
                        example: 850
                      izin:
                        type: integer
                        example: 80
                      tidak_hadir:
                        type: integer
                        example: 50
                      tamu:
                        type: integer
                        example: 20
                      attendance_rate:
                        type: number
                        format: float
                        example: 85.0
              example:
                success: true
                data:
                  total_records: 1000
                  hadir: 850
                  izin: 80
                  tidak_hadir: 50
                  tamu: 20
                  attendance_rate: 85.0
  /analytics/attendance/trends:
    get:
      tags: [Analytics]
      summary: Get attendance trends
      description: Get weekly/monthly attendance trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, monthly]
            default: weekly
        - name: cg_id
          in: query
          description: Filter by CGF ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        period:
                          type: string
                          example: "2024-W03"
                        total_expected:
                          type: integer
                          example: 50
                        total_present:
                          type: integer
                          example: 42
                        attendance_rate:
                          type: number
                          format: float
                          example: 84.0
...
  /analytics/events/summary:
    get:
      tags: [Analytics]
      summary: Get events summary
      description: Get event statistics
      parameters:
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
            example: 2024
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_events:
                        type: integer
                        example: 20
                      events_by_category:
                        type: object
                        additionalProperties:
                          type: integer
                        example:
                          Camp: 2
                          Retreat: 3
                          Quarterly: 4
                          Monthly: 8
                          Special: 3
                      total_participants:
                        type: integer
                        example: 1500
                      average_participants_per_event:
                        type: number
                        format: float
                        example: 75.0
  /analytics/members/birthday:
    get:
      tags: [Analytics]
      summary: Get upcoming birthdays
      description: Get members with birthdays in the next N days
      parameters:
        - name: days
          in: query
          description: Number of days to look ahead
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        no_jemaat:
                          type: integer
                        nama_jemaat:
                          type: string
                        tanggal_lahir:
                          type: string
                          format: date
                        upcoming_birthday:
                          type: string
                          format: date
                        age_turning:
                          type: integer
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== PARAMETERS ====================
...